Fix things from renaming and clippy

+++ src/codec/packet/side_data.rs
@@ -192,6 +192,6 @@ impl<'a> SideData<'a> {
     }
 
     pub fn data(&self) -> &[u8] {
-        unsafe { slice::from_raw_parts((*self.as_ptr()).data, (*self.as_ptr()).size as usize) }
+        unsafe { slice::from_raw_parts((*self.as_ptr()).data, (*self.as_ptr()).size) }
     }
 }
