+++ src/util/frame/side_data.rs
@@ -216,7 +216,7 @@ impl<'a> SideData<'a> {
 
     #[inline]
     pub fn data(&self) -> &[u8] {
-        unsafe { slice::from_raw_parts((*self.as_ptr()).data, (*self.as_ptr()).size as usize) }
+        unsafe { slice::from_raw_parts((*self.as_ptr()).data, (*self.as_ptr()).size) }
     }
 
     #[inline]
